+++ R/purrr-map.R
@@ -9,6 +9,7 @@
 #'
 #' @param .x A list or vector
 #' @param .y A list or vector
+#' @param .l A list to use in `pmap`
 #' @param .f A function
 #' @param ... Other arguments to pass to a function
 #' @param .id Whether `map_dfr.()` should add an id column to the finished dataset